requirements.txt
tensorflow>=2.12
hydra-core>=1.3
omegaconf>=2.3
matplotlib
pytest

